extract ObjectCrationTransformer into separate file
